separated engine class from running scripts to prevent wrong imports
